+++ crates/runtime/tests/acceleration/mod.rs
@@ -14,6 +14,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
+use spicepod::component::{dataset::acceleration::Mode, params::Params};
+
 #[cfg(feature = "duckdb")]
 mod metadata_duckdb;
 #[cfg(feature = "postgres")]
@@ -23,3 +25,15 @@ mod metadata_sqlite;
 #[cfg(all(feature = "postgres", feature = "duckdb", feature = "sqlite"))]
 mod on_conflict;
 mod query_push_down;
+
+fn get_params(mode: &Mode, file: Option<String>, engine: &str) -> Option<Params> {
+    let param_name = format!("{engine}_file",);
+    if mode == &Mode::File {
+        return Some(Params::from_string_map(
+            vec![(param_name, file.unwrap_or_default())]
+                .into_iter()
+                .collect(),
+        ));
+    }
+    None
+}
